Revert 152796 - Fixed build stopper
BUG=none
TEST=none
TBR=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10880010

TBR=skuhne@google.com
Review URL: https://chromiumcodereview.appspot.com/10878014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@152797 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 902964e24c486fc7bd05bf0873c6a68cf4f1b46b

+++ ash/wm/system_gesture_event_filter.cc
@@ -92,7 +92,7 @@ const double kAllowableScreenOverlapForDeviceCommand = 0.0005;
 const int kIgnoreFirstBezelDeviceEvents = 10;
 // Within these 'n' huge coordinate changes are not allowed. The threshold is
 // given in fraction of screen resolution changes.
-const double kBezelNoiseDeltaFilter = 0.1;
+const int kBezelNoiseDeltaFilter = 0.1;
 // To avoid the most frequent noise (extreme locations) the bezel percent
 // sliders will not cover the entire screen. We scale therefore the percent
 // value by this many percent for minima and maxima extension.
